# public_html/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your Awesome Modules's API",
        "contact": {
            "name": "Developer Team",
            "email": "developers@module.com"
        },
        "version": "2.0.0"
    },
    "paths": {
...
        "/rank-firebase": {
            "get": {
                "tags": [
                    "Firebase Rank"
                ],
                "operationId": "/rank-firebase/",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Firebase Rank"
                ],
                "summary": "Create Rank",
                "description": "Create Rank",
                "operationId": "create rank-firebase",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "5645565s4d0sd4546654",
                                        "type": "string"
                                    },
                                    "attendance_point": {
                                        "description": "5",
                                        "type": "integer"
                                    },
                                    "performance_point": {
                                        "description": "5",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/rank-firebase/{id}": {
            "get": {
                "tags": [
                    "Firebase Rank"
                ],
                "summary": "Get rank information",
                "description": "Returns rank data",
                "operationId": "getRankFirebaseById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Rank id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Firebase Rank"
                ],
                "summary": "Update Rank",
                "description": "Update Rank",
                "operationId": "update rank firebase by id",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "5645565s4d0sd4546654",
                                        "type": "string"
                                    },
                                    "attendance_point": {
                                        "description": "5",
                                        "type": "integer"
                                    },
                                    "performance_point": {
                                        "description": "5",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Firebase Rank"
                ],
                "summary": "Delete existing rank data",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteRankFirebase",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Rank id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}